Support exporting datasets and slices in local file provider
