+++ packages/h5wasm/src/local/h5wasm-local-file-api.ts
@@ -1,12 +1,16 @@
-import type { ValuesStoreParams } from '@h5web/app';
+import type { ExportFormat, ExportURL, ValuesStoreParams } from '@h5web/app';
 import { DataProviderApi } from '@h5web/app';
 import type {
+  ArrayShape,
   AttributeValues,
+  Dataset,
   Entity,
   ProvidedEntity,
+  Value,
 } from '@h5web/shared/hdf5-models';
 import type { Remote } from 'comlink';
 
+import type { Plugin } from '../models';
 import { getEnhancedError, hasBigInts, sanitizeBigInts } from '../utils';
 import { getH5WasmRemote } from './remote';
 import type { H5WasmWorkerAPI } from './worker';
@@ -15,7 +19,10 @@ export class H5WasmLocalFileApi extends DataProviderApi {
   private readonly remote: Remote<H5WasmWorkerAPI>;
   private readonly fileId: Promise<bigint>;
 
-  public constructor(file: File) {
+  public constructor(
+    file: File,
+    private readonly _getExportURL?: DataProviderApi['getExportURL'],
+  ) {
     super(file.name);
 
     this.remote = getH5WasmRemote();
@@ -51,6 +58,24 @@ export class H5WasmLocalFileApi extends DataProviderApi {
         ]),
       ),
     );
+  }
+
+  public override getExportURL<D extends Dataset<ArrayShape>>(
+    format: ExportFormat,
+    dataset: D,
+    selection: string | undefined,
+    value: Value<D>,
+  ): ExportURL {
+    const url = this._getExportURL?.(format, dataset, selection, value);
+    if (url) {
+      return url;
+    }
+
+    if (format === 'json') {
+      return async () => new Blob([JSON.stringify(value, null, 2)]);
+    }
+
+    return undefined;
   }
 
   public async cleanUp(): Promise<number> {
